[refactor] 회원 탈퇴 컴포넌트 분리
